Added different layout for products/categories based on layout-sidenav.
Different ui needed for product categories (add, edit, remove product)
Decouples sidenav layout from content layouts.
Refactored and fixed bugs for product editor app.
