--- _includes/vue-editor-app.html
@@ -1,15 +1,0 @@
-<div id="edit-app" data-title="{{ page.title }}" data-permalink="{{ page.permalink }}"
-    data-category="{{ page.category }}">
-</div>
-
-<form hidden name="edit-form" method="POST" data-netlify="true" data-netlify-honeypot="honeypot-field">
-    <!-- Spam guard -->
-    <input type="hidden" name="honeypot-field" value="edit-form" />
-    <input name="page_path" />
-    <input name="page_title" />
-    <input name="page_category" />
-    <input name="page_content" />
-    <input name="form_name" />
-</form>
-
-<script src="/assets/js/vue-editor-app.bundle.js" type="text/javascript"></script>
