+++ _layouts/layout-product.html
@@ -1,0 +1,23 @@
+---
+layout: layout-sidenav
+---
+<div class="content article">
+    {{content}}
+</div>
+<div class="has-background-light padding margin--top--lg">
+    <div id="edit-app" data-title="{{ page.title }}" data-permalink="{{ page.permalink }}"
+        data-category="{{ page.category }}">
+    </div>
+
+    <form hidden name="edit-form" method="POST" data-netlify="true" data-netlify-honeypot="honeypot-field">
+        <!-- Spam guard -->
+        <input type="hidden" name="honeypot-field" value="edit-form" />
+        <input name="page_path" />
+        <input name="page_title" />
+        <input name="page_category" />
+        <input name="page_content" />
+        <input name="form_name" />
+    </form>
+
+    <script src="/assets/js/vue-edit-product-app.bundle.js" type="text/javascript"></script>
+</div>
