+++ _layouts/layout-sidenav.html
@@ -54,14 +54,7 @@ layout: layout-default
                     </aside>
                 </div>
             </div>
-            <div class="col is-9 is-12-touch">
-                <div class="content article">
-                    {{content}}
-                </div>
-                <div class="has-background-light padding margin--top--lg">
-                    {% include vue-editor-app.html %}
-                </div>
-            </div>
+            <div class="col is-9 is-12-touch">{{content}}</div>
         </div>
     </div>
 </section>
