+++ contents/products/analytics/wogaa/index.html
@@ -1,6 +1,6 @@
 ---
 title: WOGAA
-layout: layout-sidenav
+layout: layout-product
 permalink: /products/analytics/wogaa/
 category: Products
 breadcrumb: WOGAA
